requirements.txt
django==4.1.7
djangorestframework==3.14.0
psycopg2==2.9.5
daphne==4.0.0